feat: le jardin, a broccoli-inspired rewrite
A rewrite of the transformation pipeline. The pipeline consists of a series of transformations applied to a graph. A transformation is now representable by a lettable/pipeable rxjs operator, i.e. a unary function over the build graph. Transformation functions may be composed through dependency injection, thus making the pipeline more "pluggable" and allowing to mash with custom transformations.